[FLINK-1735] change package
