+++ flink-staging/flink-ml/src/main/scala/org/apache/flink/ml/feature/extraction/FeatureHasher.scala
@@ -16,14 +16,15 @@
  * limitations under the License.
  */
 
-package org.apache.flink.ml.preprocessing
+package org.apache.flink.ml.feature.extraction
 
 import org.apache.flink.api.common.functions._
 import org.apache.flink.api.scala._
 import org.apache.flink.ml.common.{Parameter, ParameterMap, Transformer}
 import org.apache.flink.ml.math.Breeze._
 import org.apache.flink.ml.math.{DenseVector, Vector}
-import org.apache.flink.ml.preprocessing.FeatureHasher.NumberFeatures
+import org.apache.flink.ml.feature.extraction.FeatureHasher.NumberFeatures
+
 import scala.util.hashing.MurmurHash3
 
 /** Apply feature hashing on observations, so that all features have a user-specified dimensionality.
@@ -37,12 +38,12 @@ import scala.util.hashing.MurmurHash3
   * [[Vector]].
   *
   * @example
-  *          {{{
-  *            val trainingDS: DataSet[Vector] = env.fromCollection(data)
-  *            val transformer = FeatureHasher().setNumberFeatures(1000)
+  * {{{
+  *             val trainingDS: DataSet[Vector] = env.fromCollection(data)
+  *             val transformer = FeatureHasher().setNumberFeatures(1000)
   *
-  *            transformer.transform(trainingDS)
-  *          }}}
+  *             transformer.transform(trainingDS)
+  * }}}
   *
   * =Parameters=
   *
@@ -50,53 +51,57 @@ import scala.util.hashing.MurmurHash3
   */
 class FeatureHasher extends Transformer[Vector, Vector] with Serializable {
 
-  /** Sets the target number of features of the transformed data
-    *
-    * @param n the user-specified number of features.
-    * @return the FeatureHasher instance with its number of features set to the user-specified value
-    */
-  def setNumberFeatures(n: Int): FeatureHasher = {
-    parameters.add(NumberFeatures, n)
-    this
-  }
+	/** Sets the target number of features of the transformed data
+	  *
+	  * @param n the user-specified number of features.
+	  * @return the FeatureHasher instance with its number of features set to the user-specified value
+	  */
+	def setNumberFeatures(n: Int): FeatureHasher = {
+		parameters.add(NumberFeatures, n)
+		this
+	}
 
-  override def transform(input: DataSet[Vector], parameters: ParameterMap):
-  DataSet[Vector] = {
-    val resultingParameters = this.parameters ++ parameters
-    val n = resultingParameters(NumberFeatures)
+	override def transform(input: DataSet[Vector], parameters: ParameterMap):
+	DataSet[Vector] = {
+		val resultingParameters = this.parameters ++ parameters
+		val n = resultingParameters(NumberFeatures)
 
-    input.map(new RichMapFunction[Vector, Vector]() {
+		input.map(new RichMapFunction[Vector, Vector]() {
 
-      override def map(vector: Vector): Vector = {
-        var myVector = vector.asBreeze
-        
-        var hashBuckets = Array.fill[Double](n)(0.0)
-        
-        val iterator = myVector.valuesIterator
-        
-        for (i <- 0 until myVector.size) {
-          val value = iterator.next          
-          
-          val h = Math.abs(MurmurHash3.arrayHash[Double](Array(value)))
-          
-          hashBuckets(Math.abs(h) % n) += (if (h > 0) { 1.0 } else { -1.0 })          
-        }
-        
-        var hashedVector = new DenseVector(hashBuckets)
-        
-        return hashedVector
-      }
-    })
-  }
+			override def map(vector: Vector): Vector = {
+				var myVector = vector.asBreeze
+
+				var hashBuckets = Array.fill[Double](n)(0.0)
+
+				val iterator = myVector.valuesIterator
+
+				for (i <- 0 until myVector.size) {
+					val value = iterator.next
+
+					val h = Math.abs(MurmurHash3.arrayHash[Double](Array(value)))
+
+					hashBuckets(Math.abs(h) % n) += (if (h > 0) {
+						1.0
+					} else {
+						-1.0
+					})
+				}
+
+				var hashedVector = new DenseVector(hashBuckets)
+
+				return hashedVector
+			}
+		})
+	}
 }
 
 object FeatureHasher {
 
-  case object NumberFeatures extends Parameter[Int] {
-    override val defaultValue: Option[Int] = Some(1048576)
-  }
+	case object NumberFeatures extends Parameter[Int] {
+		override val defaultValue: Option[Int] = Some(1048576)
+	}
 
-  def apply(): FeatureHasher = {
-    new FeatureHasher()
-  }
+	def apply(): FeatureHasher = {
+		new FeatureHasher()
+	}
 }
